+++ fe/fe-core/src/test/java/com/starrocks/storagevolume/StorageVolumeTest.java
@@ -1,0 +1,232 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.starrocks.storagevolume;
+
+import com.staros.proto.AwsAssumeIamRoleCredentialInfo;
+import com.staros.proto.AwsSimpleCredentialInfo;
+import com.staros.proto.FileStoreInfo;
+import com.staros.proto.FileStoreType;
+import com.staros.proto.S3FileStoreInfo;
+import com.starrocks.common.AnalysisException;
+import com.starrocks.credential.CloudConfiguration;
+import com.starrocks.credential.CloudType;
+import com.starrocks.credential.aws.AWSCloudConfiguration;
+import com.starrocks.credential.hdfs.HDFSCloudConfiguration;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+
+import static com.starrocks.credential.CloudConfigurationConstants.AWS_S3_ACCESS_KEY;
+import static com.starrocks.credential.CloudConfigurationConstants.AWS_S3_ENDPOINT;
+import static com.starrocks.credential.CloudConfigurationConstants.AWS_S3_EXTERNAL_ID;
+import static com.starrocks.credential.CloudConfigurationConstants.AWS_S3_IAM_ROLE_ARN;
+import static com.starrocks.credential.CloudConfigurationConstants.AWS_S3_REGION;
+import static com.starrocks.credential.CloudConfigurationConstants.AWS_S3_SECRET_KEY;
+import static com.starrocks.credential.CloudConfigurationConstants.AWS_S3_USE_AWS_SDK_DEFAULT_BEHAVIOR;
+import static com.starrocks.credential.CloudConfigurationConstants.AWS_S3_USE_INSTANCE_PROFILE;
+
+public class StorageVolumeTest {
+    @Test
+    public void testAWSDefaultCredential() throws AnalysisException {
+        Map<String, String> storageParams = new HashMap<>();
+        storageParams.put(AWS_S3_REGION, "region");
+        storageParams.put(AWS_S3_ENDPOINT, "endpoint");
+        storageParams.put(AWS_S3_USE_AWS_SDK_DEFAULT_BEHAVIOR, "true");
+
+        StorageVolume sv = new StorageVolume("test", "s3", Arrays.asList("s3://abc"),
+                storageParams, true, "");
+        CloudConfiguration cloudConfiguration = sv.getCloudConfiguration();
+        Assert.assertEquals(CloudType.AWS, cloudConfiguration.getCloudType());
+        FileStoreInfo fileStore = cloudConfiguration.toFileStoreInfo();
+        Assert.assertEquals(FileStoreType.S3, fileStore.getFsType());
+        Assert.assertTrue(fileStore.hasS3FsInfo());
+        S3FileStoreInfo s3FileStoreInfo = fileStore.getS3FsInfo();
+        Assert.assertTrue(s3FileStoreInfo.getCredential().hasDefaultCredential());
+        Assert.assertEquals("region", ((AWSCloudConfiguration) cloudConfiguration).getAWSCloudCredential()
+                .getRegion());
+        Assert.assertEquals("endpoint", ((AWSCloudConfiguration) cloudConfiguration).getAWSCloudCredential()
+                .getEndpoint());
+    }
+
+    @Test
+    public void testAWSSimpleCredential() throws AnalysisException {
+        Map<String, String> storageParams = new HashMap<>();
+        storageParams.put(AWS_S3_REGION, "region");
+        storageParams.put(AWS_S3_ENDPOINT, "endpoint");
+        storageParams.put(AWS_S3_ACCESS_KEY, "access_key");
+        storageParams.put(AWS_S3_SECRET_KEY, "secret_key");
+        storageParams.put(AWS_S3_USE_AWS_SDK_DEFAULT_BEHAVIOR, "false");
+
+        StorageVolume sv = new StorageVolume("test", "s3", Arrays.asList("s3://abc"),
+                storageParams, true, "");
+        CloudConfiguration cloudConfiguration = sv.getCloudConfiguration();
+        Assert.assertEquals(CloudType.AWS, cloudConfiguration.getCloudType());
+        FileStoreInfo fileStore = cloudConfiguration.toFileStoreInfo();
+        Assert.assertEquals(FileStoreType.S3, fileStore.getFsType());
+        Assert.assertTrue(fileStore.hasS3FsInfo());
+        S3FileStoreInfo s3FileStoreInfo = fileStore.getS3FsInfo();
+        Assert.assertTrue(s3FileStoreInfo.getCredential().hasSimpleCredential());
+        AwsSimpleCredentialInfo simpleCredentialInfo = s3FileStoreInfo.getCredential().getSimpleCredential();
+        Assert.assertEquals("access_key", simpleCredentialInfo.getAccessKey());
+        Assert.assertEquals("secret_key", simpleCredentialInfo.getAccessKeySecret());
+        Assert.assertEquals("region", ((AWSCloudConfiguration) cloudConfiguration).getAWSCloudCredential()
+                .getRegion());
+        Assert.assertEquals("endpoint", ((AWSCloudConfiguration) cloudConfiguration).getAWSCloudCredential()
+                .getEndpoint());
+    }
+
+    @Test
+    public void testAWSInstanceProfile() throws AnalysisException {
+        Map<String, String> storageParams = new HashMap<>();
+        storageParams.put(AWS_S3_REGION, "region");
+        storageParams.put(AWS_S3_ENDPOINT, "endpoint");
+        storageParams.put(AWS_S3_USE_INSTANCE_PROFILE, "true");
+        storageParams.put(AWS_S3_USE_AWS_SDK_DEFAULT_BEHAVIOR, "false");
+
+        StorageVolume sv = new StorageVolume("test", "s3", Arrays.asList("s3://abc"),
+                storageParams, true, "");
+        CloudConfiguration cloudConfiguration = sv.getCloudConfiguration();
+        Assert.assertEquals(CloudType.AWS, cloudConfiguration.getCloudType());
+        FileStoreInfo fileStore = cloudConfiguration.toFileStoreInfo();
+        Assert.assertEquals(FileStoreType.S3, fileStore.getFsType());
+        Assert.assertTrue(fileStore.hasS3FsInfo());
+        S3FileStoreInfo s3FileStoreInfo = fileStore.getS3FsInfo();
+        Assert.assertTrue(s3FileStoreInfo.getCredential().hasProfileCredential());
+        Assert.assertEquals("region", ((AWSCloudConfiguration) cloudConfiguration).getAWSCloudCredential()
+                .getRegion());
+        Assert.assertEquals("endpoint", ((AWSCloudConfiguration) cloudConfiguration).getAWSCloudCredential()
+                .getEndpoint());
+    }
+
+    @Test
+    public void testAWSAssumeIamRole() throws AnalysisException {
+        Map<String, String> storageParams = new HashMap<>();
+        storageParams.put(AWS_S3_REGION, "region");
+        storageParams.put(AWS_S3_ENDPOINT, "endpoint");
+        storageParams.put(AWS_S3_USE_INSTANCE_PROFILE, "true");
+        storageParams.put(AWS_S3_USE_AWS_SDK_DEFAULT_BEHAVIOR, "false");
+        storageParams.put(AWS_S3_IAM_ROLE_ARN, "iam_role_arn");
+        storageParams.put(AWS_S3_EXTERNAL_ID, "iam_role_arn");
+
+        StorageVolume sv = new StorageVolume("test", "s3", Arrays.asList("s3://abc"),
+                storageParams, true, "");
+        CloudConfiguration cloudConfiguration = sv.getCloudConfiguration();
+        Assert.assertEquals(CloudType.AWS, cloudConfiguration.getCloudType());
+        FileStoreInfo fileStore = cloudConfiguration.toFileStoreInfo();
+        Assert.assertEquals(FileStoreType.S3, fileStore.getFsType());
+        Assert.assertTrue(fileStore.hasS3FsInfo());
+        S3FileStoreInfo s3FileStoreInfo = fileStore.getS3FsInfo();
+        Assert.assertTrue(s3FileStoreInfo.getCredential().hasAssumeRoleCredential());
+        AwsAssumeIamRoleCredentialInfo assumeIamRoleCredentialInfo = s3FileStoreInfo.getCredential()
+                .getAssumeRoleCredential();
+        Assert.assertEquals("iam_role_arn", assumeIamRoleCredentialInfo.getIamRoleArn());
+        Assert.assertEquals("iam_role_arn", assumeIamRoleCredentialInfo.getIamRoleArn());
+        Assert.assertEquals("region", ((AWSCloudConfiguration) cloudConfiguration).getAWSCloudCredential()
+                .getRegion());
+        Assert.assertEquals("endpoint", ((AWSCloudConfiguration) cloudConfiguration).getAWSCloudCredential()
+                .getEndpoint());
+    }
+
+    @Test
+    public void testAWSInvalidCredential() {
+        Map<String, String> storageParams = new HashMap<>();
+        storageParams.put(AWS_S3_REGION, "region");
+        storageParams.put(AWS_S3_ENDPOINT, "endpoint");
+        storageParams.put(AWS_S3_USE_INSTANCE_PROFILE, "true");
+        storageParams.put(AWS_S3_USE_AWS_SDK_DEFAULT_BEHAVIOR, "true");
+        storageParams.put(AWS_S3_IAM_ROLE_ARN, "iam_role_arn");
+        storageParams.put(AWS_S3_EXTERNAL_ID, "iam_role_arn");
+
+        try {
+            StorageVolume sv = new StorageVolume("test", "s3", Arrays.asList("s3://abc"),
+                    storageParams, true, "");
+        } catch (AnalysisException e) {
+            Assert.assertTrue(e.getMessage().contains("Storage params is not valid"));
+        }
+    }
+
+    @Test
+    public void testHDFSSimpleCredential() throws AnalysisException {
+        Map<String, String> storageParams = new HashMap<>();
+        storageParams.put("hadoop.security.authentication", "simple");
+        storageParams.put("username", "username");
+        storageParams.put("password", "password");
+        storageParams.put("dfs.nameservices", "ha_cluster");
+        storageParams.put("dfs.ha.namenodes.ha_cluster", "ha_n1,ha_n2");
+        storageParams.put("dfs.namenode.rpc-address.ha_cluster.ha_n1", "<hdfs_host>:<hdfs_port>");
+        storageParams.put("dfs.namenode.rpc-address.ha_cluster.ha_n2", "<hdfs_host>:<hdfs_port>");
+        storageParams.put("dfs.client.failover.proxy.provider",
+                "org.apache.hadoop.hdfs.server.namenode.ha.ConfiguredFailoverProxyProvider");
+
+        StorageVolume sv = new StorageVolume("test", "hdfs", Arrays.asList("hdfs://abc"),
+                storageParams, true, "");
+        CloudConfiguration cloudConfiguration = sv.getCloudConfiguration();
+        Assert.assertEquals(CloudType.HDFS, cloudConfiguration.getCloudType());
+        HDFSCloudConfiguration hdfsCloudConfiguration = (HDFSCloudConfiguration) cloudConfiguration;
+        Assert.assertEquals("simple", hdfsCloudConfiguration.getHdfsCloudCredential().getAuthentication());
+        Assert.assertEquals(5, hdfsCloudConfiguration.getHdfsCloudCredential().getHaConfigurations().size());
+
+        Map<String, String> storageParams1 = new HashMap<>();
+        storageParams1.put("hadoop.security.authentication", "simple");
+        storageParams1.put("username", "username");
+        storageParams1.put("password", "password");
+        sv = new StorageVolume("test", "hdfs", Arrays.asList("hdfs://abc"),
+                storageParams1, true, "");
+        cloudConfiguration = sv.getCloudConfiguration();
+        Assert.assertEquals(CloudType.HDFS, cloudConfiguration.getCloudType());
+        hdfsCloudConfiguration = (HDFSCloudConfiguration) cloudConfiguration;
+        Assert.assertEquals("simple", hdfsCloudConfiguration.getHdfsCloudCredential().getAuthentication());
+        Assert.assertEquals(0, hdfsCloudConfiguration.getHdfsCloudCredential().getHaConfigurations().size());
+    }
+
+    @Test
+    public void testHDFSKerberosCredential() throws AnalysisException {
+        Map<String, String> storageParams = new HashMap<>();
+        storageParams.put("hadoop.security.authentication", "kerberos");
+        storageParams.put("kerberos_principal", "nn/abc@ABC.COM");
+        storageParams.put("kerberos_keytab", "/keytab/hive.keytab");
+        storageParams.put("dfs.nameservices", "ha_cluster");
+        storageParams.put("dfs.ha.namenodes.ha_cluster", "ha_n1,ha_n2");
+        storageParams.put("dfs.namenode.rpc-address.ha_cluster.ha_n1", "<hdfs_host>:<hdfs_port>");
+        storageParams.put("dfs.namenode.rpc-address.ha_cluster.ha_n2", "<hdfs_host>:<hdfs_port>");
+        storageParams.put("dfs.client.failover.proxy.provider",
+                "org.apache.hadoop.hdfs.server.namenode.ha.ConfiguredFailoverProxyProvider");
+
+        StorageVolume sv = new StorageVolume("test", "hdfs", Arrays.asList("hdfs://abc"),
+                storageParams, true, "");
+        CloudConfiguration cloudConfiguration = sv.getCloudConfiguration();
+        Assert.assertEquals(CloudType.HDFS, cloudConfiguration.getCloudType());
+        HDFSCloudConfiguration hdfsCloudConfiguration = (HDFSCloudConfiguration) cloudConfiguration;
+        Assert.assertEquals("kerberos", hdfsCloudConfiguration.getHdfsCloudCredential().getAuthentication());
+        Assert.assertEquals(5, hdfsCloudConfiguration.getHdfsCloudCredential().getHaConfigurations().size());
+    }
+
+    @Test
+    public void testHDFSInvalidCredential() {
+        Map<String, String> storageParams = new HashMap<>();
+        storageParams.put("hadoop.security.authentication", "kerberos");
+        storageParams.put("kerberos_principal", "nn/abc@ABC.COM");
+
+        try {
+            StorageVolume sv = new StorageVolume("test", "hdfs", Arrays.asList("hdfs://abc"),
+                    storageParams, true, "");
+        } catch (AnalysisException e) {
+            Assert.assertTrue(e.getMessage().contains("Storage params is not valid"));
+        }
+    }
+}
